% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzipmed_package.R
\name{zioutlmmed}
\alias{zioutlmmed}
\title{Mediation Analysis for Zero-Inflated Count Outcomes using MZIP}
\usage{
zioutlmmed(
  outcome,
  mediator,
  exposure,
  confounder = NULL,
  X = 1,
  Xstar = 0,
  error = "Delta",
  n = 1000,
  robust = FALSE,
  zioff = NULL
)
}
\arguments{
\item{outcome}{is the zero-inflated count outcome variable}

\item{mediator}{is the continuous mediator variable, currently only 1 mediator variable is allowed}

\item{exposure}{is the primary exposure being considered, only 1 is allowed}

\item{confounder}{is a vector of confounder variables. If no confounder variables are needed then confounder is set to NULL. If more than 1 confounder is being considered then use the cbind function, e.g. cbind(var1,var2)}

\item{X}{is the theoretical value for the exposure variable to be set at. The default is to 1}

\item{Xstar}{is the theoretical value for the exposure variable to be compared to X. The default is 0, so direct, indirect, and proportion mediated values will be for a 1 unit increase in the exposure variable.}

\item{error}{='Delta' for delta method standard errors and ='Boot' for bootstrap. Default is delta method}

\item{n}{is the number of repetition if bootstrapped errors are used. Default is 1000}

\item{robust}{indicates if a robust covariance matrix should be used for MZIP in delta method derivations. Default is FALSE.}

\item{zioff}{(optional) use to specify an offset variable within the MZIP outcome model.}
}
\value{
The function will return a list of 12 elements.
LM is the linear model regressing the exposure and covariates on the continuous mediator \cr
MZIP is the results of regressing the exposure, covariates, and mediator on the outcome using the MZIP model \cr
RRNDE is the incidence rate ratio of the direct effect \cr
RRNIE is the incidence rate ratio of the indirect effect. \cr
logRRNDEse is the standard error for the log rate ratio of NDE \cr
RRNDEci is the 95\% confidence interval for the direct effect rate ratio\cr
logRRNIEse is the standard error for  the indirect effect log rate ratio \cr
RRNIEci is the 95\% confidence interval for the indirect effect rate ratio \cr
RRTE is the total effect rate ratio \cr
logRRTEse is the standard error for the total effect log rate ratio\cr
RRTECI is the confidence interval for the total effect rate ratio \cr
PM is the proportion mediated
}
\description{
This function incorporates the MZIP model into the counterfactual approach to mediation analysis
as proposed by Vanderweele when the outcome is a Zero-Inflated count variable for cases with
continuous mediators. Standard Errors for
direct and indirect effects are computed using delta method or bootstrapping. Note: This function
assumes that the outcome is continuous and all exposure, mediator, outcome, and confounder variables
have the same sample size. Binary variables must be dummy coded prior. See vignette for information
on use of the offset.
}
\examples{
    #Example using delta method
    ziout=zioutlmmed(outcome=mzipmed_data$ziY1,mediator=mzipmed_data$lmM,
                 exposure=mzipmed_data$X,confounder=cbind(mzipmed_data$C1,
                 mzipmed_data$C2),error="Delta",robust=FALSE,X=1,Xstar=0,
                 zioff=NULL)

   #Example using boostrapping, 10 iterations used for succinctness
   ziout2=zioutlmmed(outcome=mzipmed_data$ziY1,mediator=mzipmed_data$lmM,
                 exposure=mzipmed_data$X, confounder=cbind(mzipmed_data$C1,
                 mzipmed_data$C2),error="Boot",n=10,zioff=NULL)
}
